requirements.txt
amqp==5.1.1
apns2==0.5.0
appdirs==1.4.4
asgiref==3.5.2
asn1crypto==0.24.0
billiard==3.6.4.0
boto3==1.9.63
botocore==1.12.151
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15.1
cffi==1.12.3
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==2.9.2
Django==3.2.15
django-cors-headers==2.1.0
django-filter==1.0.4
django-push-notifications==1.5.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
docutils==0.14
drf-nested-routers==0.90.2
drf-yasg==1.13.0
flake8==3.4.1
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==0.4.15
gspread==5.5.0
gunicorn==19.7.1
h2==2.6.2
hpack==3.0.0
httplib2==0.20.4
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
image==1.5.32
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
kombu==5.2.4
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
oauth2client==1.5.2
oauthlib==3.2.1
openapi-codec==1.3.2
patch==1.16
pbr==4.2.0
Pillow==7.1.2
prompt-toolkit==3.0.31
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.5.0
pygubu==0.10.5
pygubu-designer==0.13
PyJWT==1.6.4
pyparsing==3.0.9
python-dateutil==2.8.0
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.15.97
ruamel.yaml.clib==0.2.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
typing==3.6.4
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
